fix: fix windows crashed due to malloc problems.
