+++ bridge/CMakeLists.txt
@@ -106,6 +106,7 @@ list(APPEND BRIDGE_SOURCE
   foundation/string_view.cc
   foundation/native_value.cc
   foundation/native_type.cc
+  foundation/dart_readable.cc
   foundation/ui_command_buffer.cc
   foundation/ui_command_strategy.cc
   polyfill/dist/polyfill.cc
